app/requirements.txt
anyio             4.9.0
certifi           2025.4.26
click             8.1.8
colorama          0.4.6
dnspython         2.7.0
ecdsa             0.19.1
exceptiongroup    1.3.0
fastapi           0.95.1
h11               0.16.0
httpcore          1.0.9
httpx             0.28.1
idna              3.10
jose              1.0.0
loguru            0.7.0
motor             3.1.2
pip               21.1.1
pyasn1            0.6.1
pydantic          1.10.7
pymongo           4.3.3
python-dotenv     1.0.0
python-jose       3.3.0
rsa               4.9.1
setuptools        56.0.0
six               1.17.0
sniffio           1.3.1
starlette         0.26.1
typing-extensions 4.13.2
uvicorn           0.22.0
win32-setctime    1.2.0